requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==3.6.2.1
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
djangorestframework==3.12.4
humanize==3.11.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
nepali-unicode-converter==1.0.3
Pillow==8.3.2
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
